Complete user stories 26 - 28

+++ app/controllers/items_controller.rb
@@ -19,8 +19,13 @@ class ItemsController<ApplicationController
 
   def create
     merchant = Merchant.find(params[:merchant_id])
-    merchant.items.create(item_params)
-    redirect_to "/merchants/#{merchant.id}/items"
+    item = merchant.items.create(item_params)
+    if item.save
+      redirect_to "/merchants/#{merchant.id}/items"
+    else
+      flash[:error] = item.errors.full_messages.to_sentence
+      redirect_to "/merchants/#{merchant.id}/items/new"
+    end
   end
 
   def edit
@@ -30,13 +35,23 @@ class ItemsController<ApplicationController
   def update
     item = Item.find(params[:id])
     item.update(item_params)
-    redirect_to "/items/#{item.id}"
+    if item.save
+      redirect_to "/items/#{item.id}"
+    else
+      flash[:error] = item.errors.full_messages.to_sentence
+      redirect_to "/items/#{item.id}/edit"
+    end
   end
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to "/items"
+    if item.orders.empty?
+      item.destroy
+      redirect_to "/items"
+    else
+      flash[:delete_item_warning] = "This item has been ordered and cannot be deleted at this time."
+      redirect_to "/items/#{item.id}"
+    end
   end
 
   private
